move dumbell fit test out of test_localisation_algorithms (with the rationale that it's not a core algorithm)
